add mime logic in a new module
